refactor(theme): move popover helper models in separate file
